Float Modulo support

+++ src/vm.cs
@@ -2045,7 +2045,7 @@ public class VM
         curr_frame.stack.Push(Val.NewNum(this, (int)l_operand._num | (int)r_operand._num));
       break;
       case Opcodes.Mod:
-        curr_frame.stack.Push(Val.NewNum(this, (int)l_operand._num % (int)r_operand._num));
+        curr_frame.stack.Push(Val.NewNum(this, l_operand._num % r_operand._num));
       break;
     }
